passwordTextField 클래스 생성, 옵저버 등록, 라벨 연동
