+++ swift-signup/swift-signup/Base.lproj/Main.storyboard
@@ -26,7 +26,7 @@
                                 <color key="textColor" systemColor="systemGreenColor"/>
                                 <nil key="highlightedColor"/>
                             </label>
-                            <textField opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="left" contentVerticalAlignment="center" borderStyle="roundedRect" placeholder="영문 소문자, 숫자, 특수기호( _, -), 5~20자" textAlignment="natural" minimumFontSize="17" translatesAutoresizingMaskIntoConstraints="NO" id="kns-LJ-zCP" customClass="idTextField" customModule="swift_signup" customModuleProvider="target">
+                            <textField opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="left" contentVerticalAlignment="center" borderStyle="roundedRect" placeholder="영문 소문자, 숫자, 특수기호( _, -), 5~20자" textAlignment="natural" minimumFontSize="17" translatesAutoresizingMaskIntoConstraints="NO" id="kns-LJ-zCP" customClass="IdTextField" customModule="swift_signup" customModuleProvider="target">
                                 <rect key="frame" x="30" y="154" width="330" height="40"/>
                                 <constraints>
                                     <constraint firstAttribute="height" constant="40" id="0as-Q7-rgj"/>
@@ -34,7 +34,7 @@
                                 <fontDescription key="fontDescription" type="system" pointSize="14"/>
                                 <textInputTraits key="textInputTraits"/>
                             </textField>
-                            <textField opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="left" contentVerticalAlignment="center" borderStyle="roundedRect" placeholder="영문 대/소문자, 숫자, 특수문자(!@#$%) 8~16자" textAlignment="natural" minimumFontSize="17" translatesAutoresizingMaskIntoConstraints="NO" id="G0Q-sK-zyQ">
+                            <textField opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="left" contentVerticalAlignment="center" borderStyle="roundedRect" placeholder="영문 대/소문자, 숫자, 특수문자(!@#$%) 8~16자" textAlignment="natural" minimumFontSize="17" translatesAutoresizingMaskIntoConstraints="NO" id="G0Q-sK-zyQ" customClass="PasswordTextField" customModule="swift_signup" customModuleProvider="target">
                                 <rect key="frame" x="30" y="254" width="330" height="40"/>
                                 <constraints>
                                     <constraint firstAttribute="height" constant="40" id="bmi-jT-Rbu"/>
@@ -128,9 +128,9 @@
                                 <nil key="textColor"/>
                                 <nil key="highlightedColor"/>
                             </label>
-                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="LQL-fd-Krx">
-                                <rect key="frame" x="30" y="294" width="330" height="21"/>
-                                <fontDescription key="fontDescription" type="system" pointSize="17"/>
+                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="LQL-fd-Krx">
+                                <rect key="frame" x="30" y="294" width="330" height="0.0"/>
+                                <fontDescription key="fontDescription" type="system" pointSize="12"/>
                                 <nil key="textColor"/>
                                 <nil key="highlightedColor"/>
                             </label>
@@ -186,6 +186,8 @@
                     <connections>
                         <outlet property="idLabel" destination="uKS-IN-ds0" id="1H8-aH-tJG"/>
                         <outlet property="idText" destination="kns-LJ-zCP" id="8pT-S5-hk6"/>
+                        <outlet property="passwordLabel" destination="LQL-fd-Krx" id="Gii-kZ-WBJ"/>
+                        <outlet property="passwordText" destination="G0Q-sK-zyQ" id="vGU-2P-NQ2"/>
                     </connections>
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="dkx-z0-nzr" sceneMemberID="firstResponder"/>
